Remove code duplication to make pylint happy
